+++ src/Projectile.java
@@ -6,7 +6,7 @@ import java.awt.*;
 public class Projectile extends Drawable {
     private int radius;
 
-    public Projectile(int radius, int xPos, int yPos, double xVel, double yVel, Color col){
+    public Projectile(int radius, double xPos, double yPos, double xVel, double yVel, Color col){
         super(xPos,yPos,xVel,yVel,col);
         this.radius = radius;
     }
